+++ packages/genesys-spark-components/src/components/beta/gux-avatar/gux-avatar-change-photo/tests/gux-avatar-change-photo.spec.ts
@@ -1,0 +1,32 @@
+jest.mock('../../../../../utils/error/log-error', () => ({
+  __esModule: true,
+  logWarn: jest.fn()
+}));
+
+import { newSpecPage } from '@test/specTestUtils';
+import { GuxAvatarChangePhoto } from '../gux-avatar-change-photo';
+import { logWarn } from '../../../../../utils/error/log-error';
+
+describe('gux-avatar', () => {
+  describe('#render', () => {
+    it(`component with valid markup`, async () => {
+      const html = `<gux-avatar-change-photo-beta>
+        <gux-avatar-beta slot="avatar" name="Conor Darcy"></gux-avatar-beta>
+      </gux-avatar-change-photo-beta>`;
+      const page = await newSpecPage({
+        components: [GuxAvatarChangePhoto],
+        html
+      });
+      expect(page.root).toMatchSnapshot();
+      expect(logWarn).not.toHaveBeenCalled();
+    });
+  });
+
+  describe('warnings', () => {
+    it(`should log warning for gux-avatar-change-photo without avatar`, async () => {
+      const html = `<gux-avatar-change-photo-beta></gux-avatar-change-photo-beta>`;
+      await newSpecPage({ components: [GuxAvatarChangePhoto], html });
+      expect(logWarn).toHaveBeenCalled();
+    });
+  });
+});