+++ packages/genesys-spark-components/svgo.config.mjs
@@ -1,0 +1,15 @@
+export default {
+  multipass: true,
+  quiet: true,
+  plugins: [
+    { name: 'removeStyleElement' },
+    { name: 'collapseGroups' },
+    { name: 'removeComments' },
+    {
+      name: 'removeAttrs',
+      params: {
+        attrs: ['style', 'fill', 'height', 'width', 'id', 'class', 'version']
+      }
+    }
+  ]
+};