+++ shared-configs/eslint-config-genesys-spark-components/typescript.mjs
@@ -1,0 +1,19 @@
+import tsEslint from 'typescript-eslint';
+import javascript from './index.mjs';
+
+export default [
+  javascript,
+  ...tsEslint.configs.recommended,
+  {
+    name: 'eslint-config-genesys-spark-components/typescript',
+    linterOptions: {
+      noInlineConfig: false,
+      reportUnusedDisableDirectives: false
+    },
+    rules: {
+      '@typescript-eslint/no-inferrable-types': 'off',
+      '@typescript-eslint/no-redundant-type-constituents': 'off',
+      '@typescript-eslint/no-unsafe-enum-comparison': 'off'
+    }
+  }
+];