+++ shared-configs/prettier-config-genesys-spark-components/eslint.config.mjs
@@ -1,0 +1,13 @@
+import genesysSparkConfig from 'eslint-config-genesys-spark-components';
+import globals from 'globals';
+
+export default [
+  ...genesysSparkConfig,
+  {
+    languageOptions: {
+      globals: {
+        ...globals.node
+      }
+    }
+  }
+];